DRA: grab a snapshot of DRA objects and plumb to ClusterSnapshot.Initialize
Having a second snapshot object inside ClusterSnapshot isn't ideal from
readability perspective, but the DRA objects can't just be tracked
inside the NodeInfos/PodInfos.

ResourceClaims can be shared between multiple Pods, so we need some global
location for them anyway. There are ResourceSlices that aren't node-local
that the snapshot still needs to pass to the DRA scheduler plugin to
ensure correct results.

Out of multiple options I tried prototyping, having a single
source-of-truth snapshot of all DRA objects that is modified during
ClusterSnapshot operations seems the cleanest. Trying to model it in a
different way always resulted in something being really confusing, or
having to synchronize a lot of state.

The Basic ClusterSnapshot can just clone the DRA snapshot on Fork(). The
Delta implementation will need something more sophisticated, but leaving
that for the end.