DRA: handle expendable pods using DRA
CA ignores Pods with priority below a cutoff, and pretends
they aren't in the cluster. If the pods have allocated ResourceClaims,
they would still block resources on a Node. So ResourceClaims owned by
expendable pods are removed from the DRA snapshot.

Predicates are now run when scheduling Pods waiting for preemption
to their nominatedNodeName. Not sure how this works if the preempted
pod is still on the Node, I suspect the filters would fail. This needs
to be tested, left a TODO.